refactor: refactor icon item
