+++ app/src/main/java/eu/tutorial/moodle/ui/component/IconsCard.kt
@@ -1,19 +1,14 @@
 package eu.tutorial.moodle.ui.component
 
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
@@ -25,8 +20,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
@@ -37,10 +30,7 @@ import eu.tutorial.moodle.data.ActivityDto
 import eu.tutorial.moodle.data.FoodDto
 import eu.tutorial.moodle.data.PeopleDto
 import eu.tutorial.moodle.data.PlaceDto
-import eu.tutorial.moodle.data.local.Emoji
 import eu.tutorial.moodle.data.local.allEmojis
-import eu.tutorial.moodle.ui.theme.poppins
-
 
 @Composable
 fun IconsComponent(
@@ -64,8 +54,7 @@ fun IconsComponent(
         ) {
             Row(
                 modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = 8.dp),
+                    .fillMaxWidth(),
                 horizontalArrangement = Arrangement.SpaceBetween
             ) {
                 Text(
@@ -84,7 +73,7 @@ fun IconsComponent(
 
             Card(
                 modifier = Modifier
-                    .height(186.dp)
+                    .height(192.dp)
                     .background(color = Color(0XFF151515)),
                 shape = RoundedCornerShape(32.dp)
             ) {
@@ -109,8 +98,6 @@ fun IconsComponent(
                     }
                 }
             }
-
-
         }
     } else{
         Column(
@@ -161,39 +148,4 @@ fun IconsComponent(
             }
         }
     }
-
 }
-
-
-@Composable
-fun AlignYourBodyElement(
-    @DrawableRes drawable: Int,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally, // 가운데 정렬
-    ) {
-        Box(
-            modifier = Modifier
-                .size(72.dp)
-                .clip(shape = RoundedCornerShape(20.dp))
-            ,
-            contentAlignment = Alignment.Center,
-        ){
-            Box(modifier = Modifier
-                .fillMaxSize()
-                .background(Color(0XFF363637))
-            )
-            Image(
-                painter = painterResource(drawable),
-                contentDescription = null,
-                contentScale = ContentScale.Crop, // 동그랗게 만듬
-                modifier = Modifier
-                    .size(44.dp)
-
-            )
-        }
-    }
-}
-
